+++ libs/cluster-tools/src/vivarium_cluster_tools/psimulate/runner.py
@@ -191,6 +191,9 @@ def main(
     max_attempts: int,
     backup_freq: int | None,
     extra_args: dict[str, Any],
+    slack_channel: str | None = None,
+    slack_tag: str | None = None,
+    mute_slack: bool = False,
 ) -> None:
     logger.debug("Validating cluster environment.")
     cluster.validate_cluster_environment()
@@ -335,6 +338,9 @@ def main(
         command_label=f"psimulate {command}",
         monitoring_url=monitoring_url,
         results_dir=str(output_paths.root),
+        slack_channel=slack_channel,
+        slack_tag=slack_tag,
+        mute_slack=mute_slack,
     )
 
     # Spit out a performance report for the workers.